(Fix): Ahora el crear una joya retorna bien el ID

+++ 9.  implementacion de bases de datos/joyeria/src/main/java/com/ospina/joyeria/modelo/entidad/Joya.java
@@ -12,7 +12,7 @@ public class Joya {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    Integer nroIdentificatorio;
+    Long nroIdentificatorio;
     String nombre;
     String material;
     Double peso;
